pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-api-spec"
version = "0.1.0"
description = "Black-box compatibility verification framework for LLM API endpoints"
readme = "docs/README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "llm-api-spec contributors"}
]
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "click>=8.0",
    "jsonschema>=4.20",
    "openapi-spec-validator>=0.7",
    "jsonpath-ng>=1.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "respx>=0.21",
    "ruff>=0.4",
]

[project.scripts]
llm-api-validate = "llm_api_spec.cli:main"

[tool.setuptools.packages.find]
include = ["llm_api_spec*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 100
